app/requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
cryptography==2.4.2
Flask==1.0.2
Flask-GraphQL==2.0.0
Flask-JWT-Extended==3.14.0
flask-marshmallow==0.9.0
graphene==2.1.3
graphene-sqlalchemy==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
more-itertools==4.3.0
openapi-spec-validator==0.2.4
pluggy==0.8.0
promise==2.2.1
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.0.2
PyYAML==3.13
requests==2.21.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.15
swagger-ui-bundle==0.0.2
urllib3==1.24.1
uWSGI==2.0.17.1
waitress==1.1.0
Werkzeug==0.14.1
